perf: precompile GPU paths, fix CPU transient dtype
Two things, bundled because both touch the precompile path and the second
depends on the first being correct.

## GPU precompile workloads

Extended #75's CPU `@compile_workload` to CUDA, ROCm, and Metal via the
package extensions. Each extension now carries its own `@setup_workload`
/ `@compile_workload` pair that exercises the steady + transient GPU
paths on a 12³ image, guarded by `*.functional()` so machines without
the GPU skip the workload entirely. `__init__` hasn't run yet at
extension-precompile time, so the workload sets `_preferred_gpu_backend`
and `_gpu_adapt` manually inside a try/finally.

Cold-start numbers for CUDA (RTX 3060, Julia 1.12.5, 32³ blob):

|                          | Before | After  |   Δ   |
| ------------------------ | -----: | -----: | ----: |
| `using CUDA; using Tortuosity` | 6.05 s | 6.88 s |  +0.8 |
| First steady solve TTFX  | 51.7 s | 6.05 s |  -88% |
| Second solve             |  35 ms |  35 ms |    ─  |
| Time-to-first-τ          |   58 s |   13 s |  -78% |

Cost side: CUDA extension precompile went from ~7 s to ~78 s once per
Tortuosity version, paid only when a CUDA device is actually present.

## Fix TransientDiffusionProblem dtype on CPU (#77)

`TransientDiffusionProblem` took a `dtype` kwarg defaulting to `Float32`
regardless of CPU/GPU, which contradicted `docs/design.md:160`:

> `SteadyDiffusionProblem` and `TransientDiffusionProblem` use `Float32`
> for all arrays when `gpu=true` and `Float64` when `gpu=false`.

Dropped the kwarg, hard-coded `T = gpu ? Float32 : Float64` to mirror
`SteadyDiffusionProblem`. The GPU workload above lands on top of the
corrected default so it's exercising the actual user path.

Side effects of the fix:

- Every CPU transient test in `test/test_transient.jl` was passing
  `dtype=Float64` as a workaround — those all become dead kwargs and
  are removed (~20 call sites).
- `test/test_gpu_e2e.jl` had three similar redundant kwargs — also
  dropped.
- The CPU precompile workload in `src/Tortuosity.jl` was passing
  `dtype=Float32`, which meant the CPU workload added in #75 was
  compiling for a dtype CPU users never hit. Dropped that too — the
  CPU workload now correctly specialises for `Float64` on the transient
  path.

Strictly an API break (kwarg removal), but package is v0.0.5 and every
in-repo caller was either ignoring the kwarg or using it as a workaround
for the bug. Closes #77.

## Verified

- `Pkg.test()`: full suite including 44 GPU e2e tests passes
- CUDA extension precompile succeeds with workload
- Measured TTFX drop above

+++ src/Tortuosity.jl
@@ -94,7 +94,7 @@ export slab_cumulative_flux
     effective_diffusivity(c, img; axis=:x)
     formation_factor(c, img; axis=:x)
 
-    prob = TransientDiffusionProblem(img; axis=:z, bc_inlet=1, bc_outlet=0, dtype=Float32)
+    prob = TransientDiffusionProblem(img; axis=:z, bc_inlet=1, bc_outlet=0)
     tsol = solve(prob, ROCK4(); saveat=0.1, tspan=(0.0, 0.2))
     flux(tsol.u, prob.D, prob.voxel_size, prob.img, prob.axis; ind=1, pore_index=prob.pore_index)
     mass_uptake(tsol.u, prob)
